Update to Java Version 21.0.2
This is a new Snake.jar. I remade it in IntelliJ with the newest version of Java. Hopefully, this will be compatible with more systems and easy to run. The file is a bit large now (9,927KB) for such a simple application. This is because in the newer versions of Java, JavaFX no longer comes with the JDK/JRE (just absolutely lovely), so I have the JavaFX files bundled in the .jar file (not all of them though, that would take 50,000KB+). I also added some small error handling. If you do not download the .txt file, it just makes one with some initial stats.